update for Marissa
I am currently watching the CSharp videos available on Moodle to understand how the program, database, and server connect and communicate. I currently lack the knowledge of how that task is completed. I have communicated my lack of understanding to my team and while we are not completely off of our time line, we are a bit behind our projected schedule.